implement StructuredProgram and Experiment classes
