+++ src/qililab/qprogram/qprogram.py
@@ -11,15 +11,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from collections import deque
 from copy import deepcopy
 from dataclasses import replace
 from typing import overload
 
-import numpy as np
-
-from qililab.qprogram.blocks import Average, Block, ForLoop, InfiniteLoop, Loop, Parallel
+from qililab.qprogram.blocks.average import Average
+from qililab.qprogram.blocks.block import Block
 from qililab.qprogram.calibration import Calibration
 from qililab.qprogram.decorators import requires_domain
 from qililab.qprogram.operations import (
@@ -40,13 +37,14 @@ from qililab.qprogram.operations import (
     Sync,
     Wait,
 )
-from qililab.qprogram.variable import Domain, FloatVariable, IntVariable, ValueSource, Variable
+from qililab.qprogram.structured_program import StructuredProgram
+from qililab.qprogram.variable import Domain
 from qililab.waveforms import IQPair, Waveform
 from qililab.yaml import yaml
 
 
 @yaml.register_class
-class QProgram:  # pylint: disable=too-many-public-methods
+class QProgram(StructuredProgram):  # pylint: disable=too-many-public-methods
     """QProgram is a hardware-agnostic pulse-level programming interface for describing quantum programs.
 
     This class provides an interface for building quantum programs,
@@ -93,50 +91,9 @@ class QProgram:  # pylint: disable=too-many-public-methods
     """
 
     def __init__(self) -> None:
+        super().__init__()
         self.qblox = self._QbloxInterface(self)
         self.quantum_machines = self._QuantumMachinesInterface(self)
-
-        self._body: Block = Block()
-        self._buses: set[str] = set()
-        self._variables: list[Variable] = []
-        self._block_stack: deque[Block] = deque([self._body])
-
-    def _append_to_block_stack(self, block: Block):
-        self._block_stack.append(block)
-
-    def _pop_from_block_stack(self):
-        return self._block_stack.pop()
-
-    @property
-    def body(self) -> Block:
-        """Get the body of the QProgram
-
-        Returns:
-            Block: The block of the body
-        """
-        return self._body
-
-    @property
-    def buses(self) -> set[str]:
-        """Get the buses of the QProgram
-
-        Returns:
-            set[str]: A set of the names of the buses
-        """
-        return self._buses
-
-    @property
-    def variables(self) -> list[Variable]:
-        """Get the variables
-
-        Returns:
-            list[Variable]: A list of variables
-        """
-        return self._variables
-
-    @property
-    def _active_block(self) -> Block:
-        return self._block_stack[-1]
 
     def has_calibrated_waveforms_or_weights(self) -> bool:
         """Checks if QProgram has named waveforms or weights. These need to be mapped before compiling to hardware-native code.
@@ -276,41 +233,6 @@ class QProgram:  # pylint: disable=too-many-public-methods
         traverse(copied_qprogram.body)
         return copied_qprogram
 
-    def block(self):
-        """Define a generic block for scoping operations.
-
-        Blocks need to open a scope.
-
-        Returns:
-            Block: The block.
-
-        Examples:
-
-            >>> with qp.block() as block:
-            >>>    # operations that shall be executed in the block
-        """
-        return QProgram._BlockContext(qprogram=self)
-
-    def parallel(self, loops: list[Loop | ForLoop]):
-        """Define a block for running multiple loops in parallel.
-
-        Blocks need to open a scope.
-
-        Examples:
-            >>> gain = qp.variable(float)
-            >>> frequency = qp.variable(float)
-            >>> with qp.parallel(loops=[ForLoop(variable=frequency, start=0, stop=100, step=10),
-                                        ForLoop(variable=gain, start=0.0, stop=1.0, step=0.1)]):
-            >>>    # operations that shall be executed in the block
-
-        Args:
-            loops (list[Loop  |  ForLoop]): The loops to run in parallel
-
-        Returns:
-            Parallel: The parallel block.
-        """
-        return QProgram._ParallelContext(qprogram=self, loops=loops)
-
     def average(self, shots: int):
         """Define an acquire loop block with averaging in real time.
 
@@ -328,65 +250,6 @@ class QProgram:  # pylint: disable=too-many-public-methods
             >>>    # operations that shall be executed in the average block
         """
         return QProgram._AverageContext(qprogram=self, shots=shots)
-
-    def infinite_loop(self):
-        """Define an infinite loop.
-
-        Blocks need to open a scope.
-
-        Examples:
-
-            >>> with qp.infinite_loop():
-            >>>    # operations that shall be executed in the infinite loop block
-
-        Returns:
-            InfiniteLoop: The infinite loop block.
-        """
-        return QProgram._InfiniteLoopContext(qprogram=self)
-
-    def loop(self, variable: Variable, values: np.ndarray):
-        """Define a loop block to iterate values over a variable.
-
-        Blocks need to open a scope.
-
-        Args:
-            variable (Variable): The variable to be affected from the loop.
-            values (np.ndarray): The values to iterate over.
-
-        Returns:
-            Loop: The loop block.
-
-        Examples:
-
-            >>> variable = qp.variable(int)
-            >>> with qp.loop(variable=variable, values=np.array(range(100))):
-            >>>    # operations that shall be executed in the loop block
-        """
-
-        return QProgram._LoopContext(qprogram=self, variable=variable, values=values)
-
-    def for_loop(self, variable: Variable, start: int | float, stop: int | float, step: int | float = 1):
-        """Define a for_loop block to iterate values over a variable.
-
-        Blocks need to open a scope.
-
-        Args:
-            variable (Variable): The variable to be affected from the loop.
-            start (int | float): The start value.
-            stop (int | float): The stop value.
-            step (int | float, optional): The step value. Defaults to 1.
-
-        Returns:
-            Loop: The loop block.
-
-        Examples:
-
-            >>> variable = qp.variable(int)
-            >>> with qp.for_loop(variable=variable, start=0, stop=100, step=5)):
-            >>>    # operations that shall be executed in the for_loop block
-        """
-
-        return QProgram._ForLoopContext(qprogram=self, variable=variable, start=start, stop=stop, step=step)
 
     @overload
     def play(self, bus: str, waveform: Waveform | IQPair) -> None:
@@ -576,115 +439,9 @@ class QProgram:  # pylint: disable=too-many-public-methods
         self._active_block.append(operation)
         self._buses.add(bus)
 
-    def variable(self, domain: Domain, type: type[int | float] | None = None):  # pylint: disable=redefined-builtin
-        """Declare a variable.
-
-        Args:
-            type (int | float): The type of the variable.
-
-        Raises:
-            NotImplementedError: If an unsupported type is provided.
-
-        Returns:
-            IntVariable | FloatVariable: The variable.
-        """
-
-        def _int_variable(domain: Domain = Domain.Scalar) -> IntVariable:
-            variable = IntVariable(domain)
-            self._variables.append(variable)
-            return variable
-
-        def _float_variable(domain: Domain = Domain.Scalar) -> FloatVariable:
-            variable = FloatVariable(domain)
-            self._variables.append(variable)
-            return variable
-
-        if domain is Domain.Scalar and type is None:
-            raise ValueError("You must specify a type in a scalar variable.")
-        if domain is not Domain.Scalar and type is not None:
-            raise ValueError("When declaring a variable of a specific domain, its type is inferred by its domain.")
-
-        if domain is Domain.Scalar:
-            if type == int:
-                return _int_variable(domain)
-            if type == float:
-                return _float_variable(domain)
-
-        if domain == Domain.Time:
-            return _int_variable(domain)
-        if domain in [Domain.Frequency, Domain.Phase, Domain.Voltage]:
-            return _float_variable(domain)
-        raise NotImplementedError
-
-    class _BlockContext:
-        def __init__(self, qprogram: "QProgram"):
-            self.qprogram = qprogram
-            self.block: Block = Block()
-
-        def __enter__(self):
-            self.qprogram._append_to_block_stack(block=self.block)
-            return self.block
-
-        def __exit__(self, exc_type, exc_value, exc_tb):
-            block = self.qprogram._pop_from_block_stack()
-            self.qprogram._active_block.append(block)
-
-    class _InfiniteLoopContext(_BlockContext):  # pylint: disable=too-few-public-methods
-        def __init__(self, qprogram: "QProgram"):  # pylint: disable=super-init-not-called
-            self.qprogram = qprogram
-            self.block: InfiniteLoop = InfiniteLoop()
-
-    class _ParallelContext(_BlockContext):  # pylint: disable=too-few-public-methods
-        def __init__(self, qprogram: "QProgram", loops: list[Loop | ForLoop]):  # pylint: disable=super-init-not-called
-            self.qprogram = qprogram
-            self.block: Parallel = Parallel(loops=loops)
-
-        def __enter__(self) -> Parallel:
-            for loop in self.block.loops:
-                loop.variable._source = ValueSource.Dependent
-                loop.variable._value = None
-            return super().__enter__()
-
-        def __exit__(self, exc_type, exc_value, exc_tb):
-            for loop in self.block.loops:
-                loop.variable._source = ValueSource.Free
-            super().__exit__(exc_type, exc_value, exc_tb)
-
-    class _LoopContext(_BlockContext):  # pylint: disable=too-few-public-methods
-        def __init__(  # pylint: disable=super-init-not-called
-            self, qprogram: "QProgram", variable: Variable, values: np.ndarray
-        ):
-            self.qprogram = qprogram
-            self.block: Loop = Loop(variable=variable, values=values)
-
-        def __enter__(self) -> Loop:
-            self.block.variable._source = ValueSource.Dependent
-            self.block.variable._value = None
-            return super().__enter__()
-
-        def __exit__(self, exc_type, exc_value, exc_tb):
-            self.block.variable._source = ValueSource.Free
-            super().__exit__(exc_type, exc_value, exc_tb)
-
-    class _ForLoopContext(_BlockContext):  # pylint: disable=too-few-public-methods
-        def __init__(  # pylint: disable=super-init-not-called
-            self, qprogram: "QProgram", variable: Variable, start: int | float, stop: int | float, step: int | float
-        ):
-            self.qprogram = qprogram
-            self.block: ForLoop = ForLoop(variable=variable, start=start, stop=stop, step=step)
-
-        def __enter__(self) -> ForLoop:
-            self.block.variable._source = ValueSource.Dependent
-            self.block.variable._value = None
-            return super().__enter__()
-
-        def __exit__(self, exc_type, exc_value, exc_tb):
-            self.block.variable._source = ValueSource.Free
-            super().__exit__(exc_type, exc_value, exc_tb)
-
-    class _AverageContext(_BlockContext):  # pylint: disable=too-few-public-methods
+    class _AverageContext(StructuredProgram._BlockContext):  # pylint: disable=too-few-public-methods
         def __init__(self, qprogram: "QProgram", shots: int):  # pylint: disable=super-init-not-called
-            self.qprogram = qprogram
+            self.structured_program = qprogram
             self.block: Average = Average(shots=shots)
 
     # pylint: disable=protected-access, too-few-public-methods
